Move to_rev() to testutils
