+++ test/statetest/statetest.hpp
@@ -13,9 +13,6 @@ namespace json = nlohmann;
 namespace evmone::test
 {
 
-/// Translates tests fork name to EVM revision
-evmc_revision to_rev(std::string_view s);
-
 struct TestMultiTransaction : state::Transaction
 {
     struct Indexes
